lots of changes

+++ new-server/app/Http/Controllers/BranchController.php
@@ -3,6 +3,8 @@
 namespace App\Http\Controllers;
 
 use App\Models\Branch;
+use App\Models\Employment;
+use App\Traits\EmployeeTrait;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\Storage;
@@ -11,6 +13,7 @@ use Illuminate\Support\Str;
 use Illuminate\Validation\Rule;
 
 class BranchController extends Controller {
+    use EmployeeTrait;
 
     public function create(Request $request) {
         $store_id = Auth::guard('stores')->user()->id;
@@ -18,7 +21,19 @@ class BranchController extends Controller {
         $rules = [
             'name' => ['required', 'string', 'max:255', Rule::unique('branches')->where('store_id', $store_id)],
             'address' => ['required', 'string', 'max:1000'],
-            'image' => ['nullable', 'image', 'mimes:jpeg,png,jpg', 'max:5012'],
+            'new_employees' => ['nullable', 'array'],
+            'new_employees.*.name' => ['required', 'string', 'max:255'],
+            'new_employees.*.email' => ['required', 'string', 'email', 'max:255', Rule::unique('employees')->where('store_id', $store_id)],
+            'new_employees.*.password' => ['required', 'string', 'min:6', 'confirmed'],
+            'new_employees.*.roles' => ['required', 'array'],
+            'new_employees.*.roles.*' => ['required', Rule::in(['manage', 'purchase', 'sale'])],
+            'new_employess.*.phone' => ['nullable', 'string', 'max:255'],
+            'new_employees.*.birthday' => ['nullable', 'date'],
+            'new_employees.*.gender' => ['nullable', 'string'],
+            'transfer_employees' => ['nullable', 'array'],
+            'transfer_employees.*.id' => ['required', Rule::exists('employees', 'id')->where('store_id', $store_id)],
+            'transfer_employees.*.roles' => ['required', 'array'],
+            'transfer_employees.*.roles.*' => ['required', Rule::in(['manage', 'purchase', 'sale'])]
         ];
 
         $validator = Validator::make($data, $rules);
@@ -30,27 +45,34 @@ class BranchController extends Controller {
             ], 400);
         }
 
-        $isExist = $request->hasFile('image');
-
-        // replace space in branch name with underscore
-        $image_name = $store_id . str_replace(' ', '_', $data['name']);
-
-        $path = $isExist
-            ? $request->file('image')->storeAs('branches', $image_name . "." . $request->file('image')->getClientOriginalExtension() )
-            : 'branches/default.jpg';
-
         $branch = Branch::create([
             'name' => $data['name'],
             'address' => $data['address'],
             'store_id' => $store_id,
-            'image' => $path,
+            'image' => 'branches/default.jpg',
+            'image_key' => Str::uuid(),
         ]);
+
+        // if there are new employees, create them
+        if (isset($data['new_employees'])) {
+            foreach ($data['new_employees'] as $employee) {
+                $employee['branch_id'] = $branch->id;
+                $this->createEmployee($store_id, $employee);
+            }
+        }
+
+        // if there are employees to be transferred, transfer them
+        if (isset($data['transfer_employees'])) {
+            foreach ($data['transfer_employees'] as $employee) {
+                $this->transferEmployee($employee->id, $branch->id, $employee->roles);
+            }
+        }
 
         return response()->json($branch);
     }
 
-    public function getBranchImage($branch_id) {
-        $branch = Branch::find($branch_id);
+    public function getBranchImage($image_key) {
+        $branch = Branch::where('image_key', $image_key)->first();
         if(!$branch) {
             return response()->json([
                 'message' => 'Branch not found.',
@@ -66,10 +88,40 @@ class BranchController extends Controller {
         return response()->file(storage_path('app' . DIRECTORY_SEPARATOR . $file_path));
     }
 
+    public function updateBranchImage(Request $request) {
+        $store_id = Auth::guard('stores')->user()->id;
+        $data = $request->all();
+        $rules = [
+            'id' => ['required', Rule::exists('branches', 'id')->where('store_id', $store_id)],
+            'image' => ['required', 'image', 'max:2048'],
+        ];
+
+        $validator = Validator::make($data, $rules);
+        if ($validator->fails()) {
+            return response()->json([
+                'message' => 'Validation failed.',
+                'errors' => $validator->errors(),
+            ], 400);
+        }
+
+        $branch = Branch::find($data['id']);
+        $image_name = $store_id . Str::uuid();
+        $path = $request->file('image')->storeAs('branches', $image_name . "." . $request->file('image')->getClientOriginalExtension());
+
+        $branch->image = $path;
+        $branch->image_key = Str::uuid();
+        $branch->save();
+        return response()->json([
+            'message' => 'Branch image updated.'
+        ]);
+    }
+
     public function getBranches(Request $request) {
         $store_id = Auth::guard('stores')->user()->id;
 
         $search_text = $request->query('search') ?? '';
+        $sort_key = $request->query('sort_key') ?? 'name';
+        $sort_order = $request->query('sort_order') ?? 'asc';
 
         // search branch by name, address
         $branches = Branch::where('store_id', $store_id)
@@ -77,6 +129,7 @@ class BranchController extends Controller {
                 $query->where('name', 'like', '%' . $search_text . '%')
                     ->orWhere('address', 'like', '%' . $search_text . '%');
             })
+            ->orderBy($sort_key, $sort_order)
             ->get();
 
         return response()->json($branches);
@@ -99,9 +152,7 @@ class BranchController extends Controller {
 
     public function update(Request $request, $branch_id) {
         $store_id = Auth::guard('stores')->user()->id;
-
         $data = $request->all();
-        error_log($branch_id);
         $data['branch_id'] = $branch_id;
         $rules = [
             'branch_id' => ['required', 'integer', Rule::exists('branches', 'id')->where('store_id', $store_id)],
@@ -121,18 +172,20 @@ class BranchController extends Controller {
 
         $branch = Branch::where('store_id', $store_id)->where('id', $branch_id)->first();
 
-        $image_exist = $request->hasFile('image');
+        $has_image = $request->hasFile('image');
 
         // delete old image if new image is uploaded
-        if ($image_exist) {
+        if ($has_image) {
             $old_image = $branch->image;
             if ($old_image != 'branches/default.jpg') {
                 Storage::delete($old_image);
             }
+            // change image key when new image is uploaded
+            $branch->image_key = Str::uuid();
         }
 
         $image_name = $store_id . Str::uuid();
-        $path = $image_exist
+        $path = $has_image
             ? $request->file('image')->storeAs('branches', $image_name . "." . $request->file('image')->getClientOriginalExtension() )
             : $branch->image;
 
@@ -153,6 +206,22 @@ class BranchController extends Controller {
                 'message' => 'Branch not found.',
             ], 404);
         }
+
+        $employments = Employment::where('branch_id', $branch_id)->where('to', null)->get();
+        // if employments is not empty, return error
+        if ($employments->isNotEmpty()) {
+            return response()->json([
+                'message' => 'Còn nhân viên làm việc tại chi nhánh này.',
+            ], 400);
+        }
+
+        $terminated_employments = Employment::where('branch_id', $branch_id)->where('to', '!=', null)->get();
+        // delete all terminated employments roles and employment
+        foreach ($terminated_employments as $employment) {
+            $employment->roles()->delete();
+            $employment->delete();
+        }
+
         $branch->delete();
 
         return response()->json($branch);
